pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "initializer"
version = "2.0.0"
description = "Modern TUI for Linux system initialization"
authors = [{name = "DeltaV235"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "rich>=13.0.0",
    "textual>=0.41.0",
    "pyyaml>=6.0.0",
    "click>=8.0.0",
    "psutil>=5.9.0",
    "distro>=1.8.0",
]

[project.scripts]
initializer = "src.main:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
    "mypy>=1.0.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true